backend/requirements.txt
affine==2.4.0
amqp==5.3.1
asgiref==3.8.1
async-timeout==5.0.1
attrs==24.2.0
autobahn==23.1.2
Automat==24.8.1
backports.zoneinfo==0.2.1
billiard==4.2.1
boto3==1.35.77
botocore==1.35.77
celery==5.4.0
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
channels_redis==4.2.1
charset-normalizer==3.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
cligj==0.7.2
constantly==23.10.4
contextily==1.5.2
contourpy==1.1.1
cryptography==44.0.0
cycler==0.12.1
daphne==4.1.2
defusedxml==0.8.0rc2
Django==4.2.17
django-cors-headers==4.4.0
django-extensions==3.2.3
django-filter==24.3
django-rest-framework==0.1.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
djoser==2.3.1
fonttools==4.55.2
geographiclib==2.0
geopy==2.4.1
haversine==2.9.0
hyperlink==21.0.0
idna==3.10
importlib_metadata==8.5.0
importlib_resources==6.4.5
incremental==24.7.2
jmespath==1.0.1
joblib==1.4.2
kiwisolver==1.4.7
kombu==5.4.2
matplotlib==3.7.5
mercantile==1.2.1
msgpack==1.1.0
numpy==1.24.4
oauthlib==3.2.2
packaging==24.2
pillow==10.4.0
prompt_toolkit==3.0.48
psycopg==3.2.3
psycopg-binary==3.2.3
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
PyJWT==2.9.0
pyOpenSSL==24.3.0
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-decouple==3.8
python3-openid==3.2.0
redis==5.2.1
requests==2.32.3
requests-oauthlib==2.0.0
s3transfer==0.10.4
service-identity==24.2.0
six==1.17.0
snuggs==1.4.7
social-auth-app-django==5.4.2
social-auth-core==4.5.4
sqlparse==0.5.2
tomli==2.2.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.2
urllib3==1.26.20
vine==5.1.0
wcwidth==0.2.13
xyzservices==2024.9.0
zipp==3.20.2
zope.interface==7.2
